Colibri.experiments (#27)
Performance changes, fixes and tests added for the NSDI paper.

* fix issues.

Eventhough the transaction does not block the DB,
the test DB created has 1 max connection set, and
blocks the second caller (tx or DB).

Fix the admitter using a different transaction than
the one the store is using, and thus not seeing
what the store had inserted.

* fix issues.

Verify we can find the segment rsv. for a given e2e reservation.
Even when returning a collection of seg. rsv., check they are non nil.
Check the segment reservation for an e2e one is active.
Better error messages in e2e admission.

* notes

* fix remove change in BUILD

* experiments segment admission

* WIP experiments e2e admission

* WIP experiments run multithreaded

* WIP experiments run multithread, back to monothread

* WIP experiments. segment admission shows performance bug

* WIP experiments e2e different curves

* WIP experiments

* fix remove superfluous map

* fix performance issue in segment admission.

We have a new DB function GetDemandsPerSource that filters the reservations
per ingress or egress, and groups them per AS ID.

* fix performance issue 2

Segment admission now runs faster; removed
a call to GetAllSegmentRsvs and added one
to GetMaxBlockedBWPerSource.

* experiments, run segment admission

* experiments cleanup

* experiments, free X axis

* rename admission/impl to admission/stateless

* experiment, get raw data

* duplicate stateless to stateful

* restore stateless

* renamed package stateful

* renamed type to StatefulAdmission

* signature of transitDemand

* create interfaces status blocked BW table

* update DB mock

* performance of availableBW improved

* rename DB method

* fix performance issue in link ratio

* cleanup DB interface and stateful admission

* changes to temporarily improve performance for NSDI paper

* cleanup stateful admission

* rewrite stateless admission without temp demands

* remove GetMaxBlockedBWPerSource

* remove GetRsvsPerSource

* rename PersistTransit

* don't use math.Round

* organize functions

* add PersistTransitAlloc

* add PersistInDemand, PersistEgDemand

* fix bug in PersistSourceState

* use a scratch pad in stateful admission.

To keep the results of the updated states, like source demand etc.,
use a structure to hold those values until the admission is
computed.
If the admission is successful, persist the state values kept
in said structure (the scratch pad).

* stateful tables fixes and UTs

* add stateful tests for avail BW. Fix bug in store

* fix glitch converting from bw to class

* fix availableBW, add tube ratio stateful UTs

* fix bug in store. Add stateful link ratio UT

* Add faster version of newSuffix()

* Remove random suffix function.

* fix build, adapt UT

* modify RaceForSuffix UT

Co-authored-by: mawyss <mawyss@ethz.ch>