add reverse_traveling flag.
When starting a down rsv. from the non core AS, the request must travel first
to the last AS (the core AS), so that the admission request will be done in
the direction of the reservation. The flag reverse_traveling indicates that
the request is just traveling to the core AS to re-request from there.