list reservations uses path type
When querying the colibri service for the list of reservations to a destination, pass along
the path type of the reservations that will match against. This allows an endhost to use sciond
to query for paths to any core AS, from those core ASes to any core AS of the other ISD, and
from them to the final destination AS.
Adapt DB.
Fix tests.