daemon returns the list of stitchable segments.
Change the RPC of the daemon listing reservations to return the list of
stitchable up, core and down segments. The endhost is then responsible of
combining (stitching) them to reach the destination.
Extend the colibri service calls to include a corresponding list of
stitchable segments. This RPC is not functioning if called with a
SCION QUIC address, i.e., from outside of the AS.