Colibri.fix up down (#32)
* add reverse_traveling flag.

When starting a down rsv. from the non core AS, the request must travel first
to the last AS (the core AS), so that the admission request will be done in
the direction of the reservation. The flag reverse_traveling indicates that
the request is just traveling to the core AS to re-request from there.

* translate reverse_traveling flag

* use the reverse_traveling flag in the store

* let keeper try to fulfill all requests even when some fail

* fix db.NextExpirationTime

* fix bug listing rsvs

* remove TestPeer

* update mocks

* update tests

* list reservations uses path type

When querying the colibri service for the list of reservations to a destination, pass along
the path type of the reservations that will match against. This allows an endhost to use sciond
to query for paths to any core AS, from those core ASes to any core AS of the other ISD, and
from them to the final destination AS.
Adapt DB.
Fix tests.

* change signature of daemon's list reservations

* cleanup

* Add expiration time to the reservation looks.

Also simplify its structure to include an optionally empty
error message, instead of using "oneof".

* cleanup

* cleanup

* daemon returns the list of stitchable segments.

Change the RPC of the daemon listing reservations to return the list of
stitchable up, core and down segments. The endhost is then responsible of
combining (stitching) them to reach the destination.
Extend the colibri service calls to include a corresponding list of
stitchable segments. This RPC is not functioning if called with a
SCION QUIC address, i.e., from outside of the AS.

* Several fixes related to down segment reservations.

Fix down segment reservation activation
The store had a wrong activation for down segment reservations.
Fix colibri_service check for TCP address in ListStitchables.
ServiceClientOperator periodically tries to resolve address of neighbors, and if all of them
were found, stores these addresses until next tick (15 minutes).