colibri integration test.
Similar to the end2end integration tests, this test will
attempt to make a reservation, send data using it, and
clean it up, for every AS pair.
When the remote AS is the same as the local one, it will do nothing.
Just added the basic skeleton for the test.