SQLITE3 transactions fix.
SQLITE3 doesn't like parallel write-transactions. It would spit out busy errors,
with an actual confusing "database is locked" message, when attempting to write with
two transactions at the same time.
The situation is patched by overriding the write functions (e.g. PersistSegRsv) at the
transaction level, and detecting if they return an sqlite3.ErrBusy.
If so, the transaction will create a new sql.Tx object, store the older one, and try again.
Added tests.
If no MaxOpenConns is specified for colibri, the default of 100 is used.