Multihomed behavior through UDP socket creation.
Via the creation of a UDP socket with destination the next hop,
those snet sockets that were created without binding them to a local
address, can now be used regularly.
There is a additional cache that allows skipping the UDP socket creation
if the address to dial to has been previously seen.
There is a recurrent refresh function running in the background that
checks that the host interfaces' local addresses didn't change, and if
they did, clears the cache.